requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
accessible-pygments==0.0.4
    # via pydata-sphinx-theme
alabaster==0.7.13
    # via sphinx
babel==2.16.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
beautifulsoup4==4.12.3
    # via pydata-sphinx-theme
boto3==1.35.47
    # via smart-open
botocore==1.35.47
    # via
    #   boto3
    #   s3transfer
certifi==2024.8.30
    # via
    #   clickhouse-connect
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via -r requirements/base.in
clickhouse-connect==0.6.23
    # via -r requirements/base.in
colorama==0.4.6
    # via sphinx-autobuild
docutils==0.17.1
    # via
    #   pydata-sphinx-theme
    #   sphinx
    #   sphinx-panels
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==8.5.0
    # via sphinx
jinja2==3.1.4
    # via sphinx
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
livereload==2.7.0
    # via sphinx-autobuild
lz4==4.3.3
    # via clickhouse-connect
markupsafe==2.1.5
    # via jinja2
packaging==24.1
    # via
    #   pydata-sphinx-theme
    #   sphinx
pydata-sphinx-theme==0.13.3
    # via sphinx-book-theme
pygments==2.18.0
    # via
    #   accessible-pygments
    #   pydata-sphinx-theme
    #   sphinx
python-dateutil==2.9.0.post0
    # via botocore
pytz==2024.2
    # via
    #   babel
    #   clickhouse-connect
pyyaml==6.0.2
    # via
    #   -r requirements/base.in
    #   sphinxcontrib-mermaid
requests==2.32.3
    # via
    #   -r requirements/base.in
    #   sphinx
    #   sphinxcontrib-images
s3transfer==0.10.3
    # via boto3
six==1.16.0
    # via python-dateutil
smart-open[s3]==7.0.5
    # via -r requirements/base.in
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.6
    # via beautifulsoup4
sphinx==4.5.0
    # via
    #   -r requirements/base.in
    #   pydata-sphinx-theme
    #   sphinx-autobuild
    #   sphinx-book-theme
    #   sphinx-copybutton
    #   sphinx-panels
    #   sphinxcontrib-contentui
    #   sphinxcontrib-images
    #   sphinxcontrib-mermaid
    #   sphinxcontrib-yt
sphinx-autobuild==2021.3.14
    # via -r requirements/base.in
sphinx-book-theme==1.0.1
    # via -r requirements/base.in
sphinx-copybutton==0.5.2
    # via -r requirements/base.in
sphinx-panels==0.6.0
    # via -r requirements/base.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-contentui==0.2.5
    # via -r requirements/base.in
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-images==0.9.4
    # via -r requirements/base.in
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-mermaid==1.0.0
    # via -r requirements/base.in
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sphinxcontrib-yt==0.2.2
    # via -r requirements/base.in
tornado==6.4.1
    # via livereload
typing-extensions==4.12.2
    # via pydata-sphinx-theme
urllib3==1.26.20
    # via
    #   botocore
    #   clickhouse-connect
    #   requests
wrapt==1.16.0
    # via smart-open
zipp==3.20.2
    # via importlib-metadata
zstandard==0.23.0
    # via clickhouse-connect
